requirements.txt
PyPDF2==1.26.0
PyYAML==3.11
invoke==0.13.0
Flask==0.11.1
requests==2.10.0
requests-oauthlib==0.6.2
pytest==2.9.2
pytest-cov==2.3.1
hacking==0.11.0
pylint==1.6.4
flake8==2.5.5
autopep8==1.2.4
pdoc==0.3.2
